+++ Assets/ProjectQQ/Scripts/Table/MonsterSpawnData.cs
@@ -1,8 +1,8 @@
 public struct MonsterSpawnData : IData
 {
     public int id;   // 웨이브 아이디(100101 => 앞자리부터 10 챕터, 01 스테이지, 01 웨이브 번호)
-    public int spawnStart;  // spawn 시작 시간
-    public int spawnLoop;  // spawn 반복 시간
+    public float spawnStart;  // spawn 시작 시간
+    public float spawnLoop;  // spawn 반복 시간
     public int spawnCount;  // spawn 횟수
     /// <summary>
     /// 몬스터 테이블 참조
